build: native Windows development builds with MSYS2 clang64
Development builds only: the daemons, modeld, pandad, v4l and socketcan
stay Linux. SCons runs build commands through MSYS2 bash so the
SConscripts' POSIX shell syntax keeps working. Portable forms replace
dirent, readlink, socketpair, aligned_alloc and the rm -rf shell-outs on
every platform; what the CRT still spells differently is shimmed in
common/util.h. replace_file retries the rename a Windows reader blocks.
teleoprtc has no Windows wheels yet, so it is marked off there.

+++ openpilot/tools/cabana/streams/pandastream.cc
@@ -60,7 +60,7 @@ void PandaStream::streamThread() {
     MessageBuilder msg;
     auto evt = msg.initEvent();
     auto canData = evt.initCan(raw_can_data.size());
-    for (uint i = 0; i<raw_can_data.size(); i++) {
+    for (size_t i = 0; i<raw_can_data.size(); i++) {
       canData[i].setAddress(raw_can_data[i].address);
       canData[i].setDat(kj::arrayPtr((uint8_t*)raw_can_data[i].dat.data(), raw_can_data[i].dat.size()));
       canData[i].setSrc(raw_can_data[i].src);
